requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.109.1
requests==2.32.5
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
langchain==0.3.27
langchain-openai==0.3.33
langchain-core==0.3.76
langgraph==0.6.7
gunicorn==23.0.0
